Fixed non-compiling code and warnings
git-svn-id: https://openlierox.svn.sourceforge.net/svnroot/openlierox@1788 34602234-ff1f-0410-a465-ea8f3b77ab7f

+++ src/common/CWorm_SendRecv.cpp
@@ -472,7 +472,7 @@ void CWorm::readPacketState(CBytestream *bs, CWorm *worms)
 	bs->read2Int12( x, y );
 
 	// Angle
-	fAngle = tState.iAngle = bs->readInt(1) - 90;
+	fAngle = tState.iAngle = (float)(bs->readInt(1) - 90);
 	
 	// Flags
 	uchar bits = bs->readByte();
